Don't require that active choice stays on same scheduler thread
It should now be possible to run arbitrary code inside of the
gpu-branch of a choice, including cpu-parallel code. (Exception:
choices inside of active choices are not permitted, but we don't
check for this yet.)

The way we do this is by removing the "identity" of the gpu manager,
and instead just keep track of which MLton thread is currently
executing the active choice (i.e., an active gpu-branch). This allows
the thread to be scheduled as normal, e.g., it can be safely migrated
across processors with no trouble. Whoever finishes the active choice
then is responsible for searching for a new pending choice to put
onto the GPU.