Ensure decheckSyncDepths end up on new thread correctly
Perhaps this went wrong in the git merge?

Here's the gist of the problem and the fix:
  - At promotion, the `decheckSyncDepths` of the source
    thread get copied into the new thread (good).
  - The new thread is not runnable yet. For example, it
    does not yet have its own heap---it's really just a
    single stack frame. This is somewhat important for
    performance (avoid the cost of creating a heap if
    it's never stolen).
  - When the new thread gets stolen, we copy it again.
    This time, we create a _real_ thread, with a heap to
    go along with it.
  - This is where the problem was: at the steal, the
    `decheckSyncDepths` of the real thread were being
    initialized to 0, instead of being copied from the
    intermediate thread.
  - The fix was straightforward: just need to copy the
    `decheckSyncDepths` inside of `copyThreadWithHeap`
    in the run-time system.

This patch also just does a simplifying thing: at steal, we
annotate the chunks of the _real_ new thread (the runnable one)
with BOGUS_DECHECK_TID. The only time this should matter is
in the scheduler, at the join-point synchronization, where
the right-hand side writes down its own thread and the
left-hand side then acquires this thread. We know this is not
entangled, so... let's just not check it at all.