revive special-case code for eager forking (bugfix and perf benefit?)
I believe commit 378bfa1 introduced a subtle bug... a bad interaction
between (a) clearing suspects in parallel, and (b) CGC.

This commit fixes the bug (AFAICT), and appears to have performance
benefits as well.

I don't have a complete grasp on what the bug was, but I believe it
went something like this:
  - At a join point, we merge threads and then enter into the
    `maybeParClearSuspects...` call
  - This call starts by forking, and the fork spawns a CGC
  - While clearing suspects, the CGC simultaneously reclaims
    objects that are referenced by the suspect set!
  - The result is a dangling pointer; specifically, when clearing
    suspects, we get an error where the suspect objptr points to
    reclaimed bits.

To fix this, I changed `maybeParClearSuspects...` so that it cannot
spawn a CGC. Specifically, I revived the code for doing eager forks,
and created an option for eager forking while disallowing CGC.

After making this change, it occurred to me that the specialized
eager forking code is likely to be more efficient than the previous
method of eager forking. (Previously, to do an eager fork, we would
do a pcall and then immediately trigger a promotion.)

So, I modified `par` to drop into the specialized eager forking
code in the case where a promotion token is available. This seems
to have a mild performanc benefit in some cases.