bugfix: try a different approach for assigning decheck-id's to chunks of new threads
This has been a pain point: what decheck-id should we assign to the
initial chunks for a new runnable thread? These chunks store the thread
object itself, and the call-stack of that thread. We store them in a
pseudoheap immediately above the depth where the thread begins execution.

In 22152e6ae20213d10a49fd61dc8bcb3f93014f36, I attempted a simplification:
just assign these chunks a bogus id. However, this has a problem: if we
ever return to these chunks to continue allocations, then the new allocations
will implicitly be assigned the bogus id; this in turn trips the entanglement
management system, and we get a crash.

So, in this patch, I'm trying a new approach which seems more robust. The
idea is to assign these chunks the same decheck-id as the parent task.
This makes sense in the hierarchical heap structure: because these chunks
are in a pseudoheap of the parent, it makes sense that they should have
a decheck-id which matches the parent.

AFAICT, this works. We'll see how it holds up moving forward.