decrease max capacity of the work-stealing deques
This is just a small thing that has bugged me for a while. We
have a max fork-depth constraint due to the current implementation
of the DePa-based decheck algorithm. The max fork depth is 31.
But, the deques have had a capacity of ~1000 for years, which is
just wasted space. This patch just brings the deque capacity
down to 64.