bugfix? LGC should never clear a suspect...
Previously, LGC was attempting to unpin objects and then unmark them
as suspects. This is not correct -- it's possible for a suspect to
be in scope of LGC, and it should remain a suspect, because LGC can
have one or more ancestor heaps in scope. (So, for example, an
ancestor object that contains a down-pointer would be marked as
suspect, could be in-scope of the LGC, and should remain a suspect.)