fix long-standing GC bug??
This patch appears to fix #156. The solution is different than what was
proposed in the discussion there, and also different from what was
proposed in #178. The idea here is to restrict what range of the
work-stealing deque is accessed by LGC, to ensure that the only slots
accessed are those that are in-scope of the LGC. To implement this,
I created a new `foreachObjptrInSequenceSlice` function which traces
the roots within an index range of a sequence object.