Reorder fields in `struct GC_thread` for alignment purposes
Upstream MLton added a check that object components are sufficiently
aligned (see 519aacb6); in particular, when `-align 8` (i.e., the beginning
of objects are 8-byte aligned), then 64-bit components (e.g., `Word64` or
`Real64` or `Objptr` or `CPointer`) are also 8-byte aligned.  MLton's
`PackedRepresentation` strategy establishes this for most objects, but the
`Normal` object representing a `struct GC_thread` is manually constructed.

MaPLe's current `struct GC_thread` does not satisfy this property.  It
currently looks like:

    typedef struct GC_thread {

      /* unspent heartbeats */
      uint32_t spareHeartbeats;

      int32_t currentProcNum; /* the worker currently executing this thread */
      size_t bytesNeeded;
      ptrdiff_t exnStack;

      /* Current depth (fork depth) of the thread allocating in this heap.
       * Fresh chunks are placed at this level. */
      uint32_t currentDepth;

    #ifdef DETECT_ENTANGLEMENT
      decheck_tid_t decheckState;
      uint32_t decheckSyncDepths[DECHECK_DEPTHS_LEN];
    #endif

      uint32_t minLocalCollectionDepth;

      size_t bytesAllocatedSinceLastCollection;
      size_t bytesSurvivedLastCollection;

      struct HM_HierarchicalHeap* hierarchicalHeap;

      /* The "current" chunk of the heap, where the frontier is pointing */
      HM_chunk currentChunk;

      objptr stack;
    } __attribute__ ((packed)) *GC_thread;

The `decheck_tid_t decheckState` field (which is a `union` including a
`uint64_t`) is not 8-byte aligned, which triggers the RSSA type checking
error.  This is not actually a problem for the generated code, since no
generated code attempts to access the fields of a thread.

Nonetheless, since the order of the `struct GC_thread` fields does not
matter, we simply move `uint32_t minLocalCollectionDepth` field to before
the `decheck_tid_t decheckState` field.