Allow for multiple device indices to refer to the same device string
This change decouples device indices from device strings, allowing
for two different device indices to map to the same physical device,
if desired. The motivation is to allow for us to run multiple GPU
tasks concurrently on the same device; we can do so by creating
multiple "virtual" device ids that all have the same physical id.

For example, running with
  -devices "#0,#0"
will create two virtual device ids, both of which map to the same
physical device (#0).

The next step is to adjust the scheduling algorithm to incorporate
device feedback, to prevent each physical device from becoming
overloaded. (At the moment, the scheduler assumes that each virtual
device index maps to an uncontended device, but this assumption
obviously doesn't hold if multiple virtual ids map to the same
physical device.)

By the way, the previous behavior of the scheduler is still preserved.
Previously the scheduler assumed that each element of -devices "..."
is unique, and this patch doesn't change the behavior of the scheduler
in that case.