Store PCall alternate return addresses in frame info
As suggested by @JohnReppy at POPL'24, one can exploit the static
relationship among the return addresses at a PCall to avoid storing the
alternate return addresses on the (dynamic) call stack; rather, the
`cont`/`seq` return address on the stack should be sufficient to determine
the corresponding `parl`/`sync` and `parr`/`spwn` return addresses.

In MLton, we can use the `GC_frameInfo` that is associated with each return
address to store the `parl`/`sync` and `parr`/`spwn` return addresses in a
`PCALL_CONT_FRAME`.

One advantage is that it avoids two writes (of the `parl`/`sync` and
`parr`/`spwn` return addresses) at each `PCall`.  Although it is expected
that this overhead is small (the top of the call stack ought to be in
cache), it does nicely reduce the overhead of `PCall` to exactly the same
as `Call`.