allow for "youngest frame" optimization in forkThreadAndSetData
Introduce a new prim:
  PCall_forkThreadAndSetData_youngest

This prim is the same as the normal forkThreadAndSetData, except that
it only walks the stack to find the youngest frame. The intent is that
this can be used as an optimization in the scheduler: if we have a spare
heartbeat token at the moment that we execute a pcall, then we know that
no other ancestor frames in the stack are promotable (otherwise the
spare token would have already been spent to promote them). Therefore,
we can promote the "current" pcall by only searching for the youngest
frame.

It's important to note that this is just an optimization, and doesn't
change the promotion policy. The promotion policy is still oldest-first.
The optimization is just for the case that the youngest frame happens
to coincide with the oldest.