Fix bug in order of `sporkSpwns` for `sporkNest` in `SporkNesting`
When calculating the `spork` nesting, a `Label.t list` is used, where
the head of the list is the innermost `spork` and the last element of
the list is the outermost `spork`.  But, when recording the `spork`
nesting for a block label as a `Label.t vector`, the first element
should be the outermost `spork` and the last element should be the
innermost `spork`.  This also corresponds to the "index" assigned to
each `SporkSpwn` label.