Remove "skipped heartbeat" optimization---this is not valid
Recall the desired invariant for heartbeat tokens:
  If the number of tokens >= 1, then no ancestors are promotable

This invariant is _almost always_ upheld, but we cannot guarantee
it at every moment, due to concurrency issues with the signal handler.
The heartbeat handler restores this invariant.

(Previously, we thought that the invariant was properly upheld; if
this were the case, then it's possible to "skip" heartbeats that
were triggered when we had at least one token. In this case, the
invariant would immediately tell us that no ancestors are promotable,
so no need to try to do promotions at these beats.)