Fix subtle GC bug in GC_HH_forkThread: dangling pointer
Previously, `GC_HH_forkThread` did the following steps:
  (1) find the promotable frame
  (2) populate the data slot
  (3) mark the promotable frame as PCALL_PARL_FRAME
  (4) allocate a new thread and stack
  (5) copy the promotable frame to the new stack
  (6) mark the copied frame as PCALL_PARR_FRAME

Unfortunately, this is subtly incorrect.

First, note that step 4 might trigger a GC, and the previous code was
careful to work around this by populating the data slot before the GC.
This seemed good enough -- if the GC was triggered and the data contents
were forwarded to a new heap, then by putting this data into the frame,
we guarantee it gets implicitly forwarded correctly.

However, there is a very subtle bug here. By marking the frame as a
PCALL_PARL_FRAME in step 3, we accidentally allow the garbage collector
to reclaim live data! Specifically, if there is any data which is live
in the PCALL_PARR_FRAME but not live in the PCALL_PARL_FRAME, then we'll
incorrectly reclaim this data, causing the PCALL_PARR_FRAME to go wrong.

This patch fixes the issue by doing the steps in this order:
  (1) allocate a new thread and stack
  (2) find the promotable frame
  (3) populate the data slot
  (4) copy the promotable frame to the new stack
  (5) mark the promotable frame as PCALL_PARL_FRAME
  (6) mark the copied frame as PCALL_PARR_FRAME

To do it in this order, we need to work around the possible GC in the first
step. The challenge was just that there was no good place to keep track of
the data pointer, which might be implicitly forwarded during the GC. So, I
created a new GC_state field `s->savedAdditionalRoot` which can be used to
store one additional object, and will be updated by the GC if necessary.