Introduce more loops when inlining
When inlining potentially recursive functions, we would like to convert a tail
call of a previously inlined function into a goto to the beginning of the
inlined code.  However, this tail call is nested inside the non-tail call of a
different function, then the tail call cannot be converted to a goto.

Previously, we tracked only one `return: Return.t` that corresponded to the
"return context" through all of the function contexts (the "root" function and
all of the nested inlined functions).  If this were `Tail`, then a new tail call
to a function in the context could be converted to a goto.  However, if this
were `NonTail`, then we unconditionally retained the call (with the composed
return context).  This misses some opportunities, because if the `NonTail`-ness
is "above" the occurrence of this function in the context, then the call *can*
be converted to a goto.

Now, we track one `Return.t` per function in the context.  If a call is tail and
all of the `Return.t`s up to (but not including) a match of the called function
in the function context are `Tail`, then it is converted to a goto.

See MPLLang/mpl#203