Finish spork with exception continuations
During ClosureConvert, the higher-order `spork` primitive is
elaborated using a new `DirectExp.try` ("Exceptional Syntax") form;
essentially,

  spork (body, bodyarg, spwn, spwnarg, seq, sync, exnseq, exnsync)

is expanded to a `Spork` transfer with a `cont` continuation and a
`spwn` continuation.  The `spwn` continuation is as before (evaluate
`spwn` applied to `spwnarg`, wrapped so that either a normal return or
an exceptional raise is reported as a `Bug`).  The `cont` continuation
is now more involved, the `cont` continuation is a `DirectExp.try` of
`body` applied to `bodyarg`, with a normal continuation and an
exception continuation.  In the normal continuation, a `spoin` is
executed, dispatching to either the `seq` or `sync` continuation; in
the exception continuation, a `spoin` is also executed, dispatching to
either the `exnseq` or `exnsync` continuation.  The `seq`, `sync`,
`exnseq`, and `exnsync` continuations all join control-flow to yield
the result of the `spork` primapp (or, if any of these continuations
raise an exception (as typically will the `exnseq` and `exnsync`
continuations), transfers to the handler at the point of the `spork`
primapp).