Implement promotion stacks
This commit implements promotion stacks: an auxiliary data structure
alongside every call-stack which keeps track of all of the promotable
frames within that call-stack. This allows us, at each promotion, to
find the oldest promotable frame in O(1) time. Previously, to find the
oldest promotable frame, we had to walk the stack, costing O(#frames).

Promotion stacks are implemented using three additional values:
```c
  struct GC_stack {
    ...
    size_t promoStackReserved; // number of bytes reserved for the promotion stack
    pointer promoStackBot;
    pointer promoStackTop;
    ...
  }
```

These keep track of the bottom and top of the promotion stack:

                      reserved bytes    promoStackReserved bytes
                            |                |
  +----------+--------+-----------+----------------+
  | metadata | struct | frames... | promo stack... |
  +----------+--------+-----------+----------------+
                                     ^   ^
                                     |   |
                          promoStackBot  |
                                        promoStackTop

Each element of the promo stack is a pointer to a promotable frame.
The bottommost element is a pointer to the oldest promotable frame;
the topmost element is the youngest. Stacks therefore have to be
specially handled by LGC: whenever a stack is copied, we have to
update the elements of the promo stack.

Note that promoStackTop is exclusive of the topmost element (i.e., the
promo stack is empty whenever promoStackBot == promoStackTop).

These values are maintained in the compiled code as follows:

  At every PCall:
    *(pointer*)promoStackTop = StackTop
    promoStackTop += sizeof(pointer)
  At every seq (unpromoted) continuation:
    promoStackTop -= sizeof(pointer)
  At every sync (promoted) continuation:
    promoStackTop -= sizeof(pointer)
    promoStackBot = promoStackTop
  At promotion:
    oldestPromotableFrame = *(pointer*)promoStackBot;
    promoStackBot += sizeof(pointer)
    ... promote oldestPromotableFrame ...

This requires only a few extra instructions on the fast path. With
this update, each PCall is slightly more expensive. The cost is very
small: similar to passing just one extra argument on the call-stack
at every PCall.