update ForkJoin managed loops to choose efficient loop index representation based on -default-type int<N>
Previously, the managed loops inside ForkJoin all used Word64 as
an internal loop index representation. The goal is to skip unnecessary
overflow checks, which improves performance slightly.

This patch changes it so that, if `-default-type int<N>` is
selected at compile time, then the ForkJoin managed loops will use
Word<N>.

If the default int type is set to intinf, then IntInf will be used
for the managed loops; this will have overflow checks, but the
overheads of using IntInf are huge anyway, so no big deal.

The implementation introduces a new functor, Int_ChooseFromInt,
similar to the Int_ChooseInt functor, except that it allows for
the chosen function to be polymorphic, which we need for the
type signatures of the managed loops.