re-enable scheduler thread local collections
This was a space leak. These local collections appear
to have been disabled as part of the entanglement
management effort a few years ago (specifically, in
commit 4b4d1ce, which was then carried through to commit
51c658c more recently during the PCall implementation
effort).

Scheduler threads live at hierarchical heap depth 1, with
the "global" heap immediately above as a shared ancestor.
Interpreting this whole system as a disentangled heap
hierarchy, we can perform scheduler-thread-local GC in
the same way we would a leaf collection.